requirements.txt
Flask
pandas
openpyxl

# LangChain + Google GenAI
langchain
langgraph
langchain-google-genai
google-generativeai
langchain_classic
langchain_community

# Core langchain runtime used by other adapters (provides langchain_core)
langchain-core

# Pydantic v1 required by some langchain integrations (avoid v2 runtime errors)
pydantic


# DB + Sessions
flask_sqlalchemy
Flask-Session

# Utilities
gunicorn
python-dotenv
requests
beautifulsoup4
cloudscraper
chardet

# ML
scikit-learn

# Optional tools
duckduckgo-search
ddgs

psycopg2-binary 
sqlalchemy 
flask-sqlalchemy

tavily-python
matplotlib 
fpdf

langchain_groq
flask_login
asgiref
prophet